Support exploring ".wat" files
Convert parseWasm() to be a global function instead of an an instance
method of the Run command. This wraps WasmKit.parseWasm() and supports
both ".wat" and ".wasm" files.
