+++ Sources/CLI/Commands/Run.swift
@@ -1,6 +1,5 @@
 import ArgumentParser
 import SystemPackage
-import WAT
 import WasmKit
 import WasmKitWASI
 
@@ -120,12 +119,12 @@ struct Run: ParsableCommand {
         let module: Module
         if verbose, #available(macOS 13.0, iOS 16.0, watchOS 9.0, tvOS 16.0, *) {
             let (parsedModule, parseTime) = try measure {
-                try self.parseWasm(filePath: FilePath(path))
+                try parseWasm(filePath: FilePath(path))
             }
             log("Finished parsing module: \(parseTime)", verbose: true)
             module = parsedModule
         } else {
-            module = try self.parseWasm(filePath: FilePath(path))
+            module = try parseWasm(filePath: FilePath(path))
         }
 
         let (interceptor, finalize) = try deriveInterceptor()
@@ -146,23 +145,6 @@ struct Run: ParsableCommand {
             log("Finished invoking function \"\(path)\": \(invokeTime)", verbose: true)
         } else {
             try invoke()
-        }
-    }
-
-    /// Parses a `.wasm` or `.wat` module.
-    func parseWasm(filePath: FilePath) throws -> Module {
-        if filePath.extension == "wat", #available(macOS 11.0, iOS 14.0, macCatalyst 14.0, tvOS 14.0, visionOS 1.0, watchOS 7.0, *) {
-            let fileHandle = try FileDescriptor.open(filePath, .readOnly)
-            defer { try? fileHandle.close() }
-
-            let size = try fileHandle.seek(offset: 0, from: .end)
-
-            let wat = try String(unsafeUninitializedCapacity: Int(size)) {
-                try fileHandle.read(fromAbsoluteOffset: 0, into: .init($0))
-            }
-            return try WasmKit.parseWasm(bytes: wat2wasm(wat))
-        } else {
-            return try WasmKit.parseWasm(filePath: filePath)
         }
     }
